# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /patients:
    get:
      operationId: listPatients
      tags: [patients]
      summary: List patients
      description: Returns the patients accessible to the signed-in provider.
      responses:
        "200":
          description: A list of patients
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Patient"
    post:
      operationId: createPatient
      tags: [patients]
      summary: Onboard a new patient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePatientRequest"
      responses:
        "201":
          description: Patient created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Patient"
        "409":
          description: A patient with this MRN already exists
  /patients/sync:
    post:
      operationId: syncPatientFromEhr
      tags: [patients]
      summary: Pull a patient from the configured EHR
      description: >-
        Reads `Patient/{externalId}` from the EHR named by `EHR_MODE`
        and upserts the local row keyed on MRN. Returns 201 when a new
        row was inserted and 200 when an existing row was refreshed.
        Mock mode (EHR_MODE unset) synthesizes demographics from the id
        for local development.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncPatientRequest"
      responses:
        "200":
          description: Existing patient refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncedPatient"
        "201":
          description: New patient inserted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncedPatient"
        "400":
          description: External id missing or empty
        "404":
          description: The EHR returned no patient for this external id
        "422":
          description: The EHR record was missing fields HaloNote requires
        "502":
          description: The EHR returned an error
...
  /schedule/today:
    get:
      operationId: getTodaySchedule
      tags: [schedule]
      summary: Appointments for the signed-in provider on a given day
      description: >-
        Returns appointments from the configured EHR for the requested
        local-date (defaults to today), scoped to the calling user's
        `ehrPractitionerId`. Returns 409 if the user is not linked to
        an EHR practitioner yet. Mock mode returns a synthetic roster
        when EHR_MODE is unset.
      parameters:
        - in: query
          name: date
          required: false
          schema:
            type: string
            pattern: "^\\d{4}-\\d{2}-\\d{2}$"
          description: >-
            Local-date in YYYY-MM-DD. If omitted, today's date is used.
      responses:
        "200":
          description: Schedule
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/ScheduledAppointment"
        "400":
          description: date query param is not in YYYY-MM-DD format
        "409":
          description: Caller has no ehrPractitionerId set
        "502":
          description: EHR returned an error
  /patients/{id}/history:
    get:
      operationId: getPatientHistory
      tags: [patients]
      summary: Active problems, medications, and allergies for a patient
      description: >-
        Pulls Condition (active), MedicationRequest (active), and
        AllergyIntolerance resources from the configured EHR in parallel
        and shapes them for a note-prep panel. The patient id parameter
        is the EHR Patient.id (same value used for /patients/sync).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Patient history snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PatientHistory"
        "502":
          description: EHR returned an error
...
components:
  schemas:
...
    Patient:
      type: object
      required: [id, firstName, lastName, dateOfBirth, mrn]
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        # ISO 8601 date, e.g. "1985-04-12"
        dateOfBirth:
          type: string
        mrn:
          type: string
          description: Medical record number
...
    CreatePatientRequest:
      type: object
      required: [firstName, lastName, dateOfBirth, mrn]
      properties:
        firstName:
          type: string
          minLength: 1
        lastName:
          type: string
          minLength: 1
        dateOfBirth:
          type: string
          description: ISO 8601 date, e.g. "1985-04-12"
          pattern: "^\\d{4}-\\d{2}-\\d{2}$"
        mrn:
          type: string
          minLength: 1
          description: Medical record number. Must be unique.
    SyncPatientRequest:
      type: object
      required: [externalId]
      properties:
        externalId:
          type: string
          minLength: 1
          description: The EHR-side patient id to read (e.g. FHIR Patient.id).
    SyncedPatient:
      type: object
      required: [id, firstName, lastName, dateOfBirth, mrn, synced]
      properties:
        id:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
        mrn:
          type: string
        synced:
          type: object
          required: [provider, created]
          properties:
            provider:
              type: string
              enum: [athenahealth, epic, mock]
            created:
              type: boolean
              description: True when this call inserted a new row; false when it refreshed.
...
    ScheduledAppointment:
      type: object
      required: [appointmentId, start, end, status, reason, patient]
      properties:
        appointmentId:
          type: string
        start:
          type: string
          format: date-time
        end:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
        reason:
          type: ["string", "null"]
        patient:
          oneOf:
            - type: "null"
            - type: object
              required: [ehrId, display]
              properties:
                ehrId:
                  type: string
                display:
                  type: string
    PatientHistory:
      type: object
      required: [problems, medications, allergies]
      properties:
        problems:
          type: array
          items:
            type: object
            required: [id, text, onsetDate]
            properties:
              id:
                type: string
              text:
                type: string
              onsetDate:
                type: ["string", "null"]
        medications:
          type: array
          items:
            type: object
            required: [id, text, dosage]
            properties:
              id:
                type: string
              text:
                type: string
              dosage:
                type: ["string", "null"]
        allergies:
          type: array
          items:
            type: object
            required: [id, text, severity, reactions]
            properties:
              id:
                type: string
              text:
                type: string
              severity:
                type: ["string", "null"]
              reactions:
                type: array
                items:
                  type: string
...